TW-541: Refactor `UploadFileApi` tp `MediaApi`
