+++ lib/data/network/extensions/file_info_extension.dart
